notifier slack: prevent swallowing errors
The notifier functions are set up to return errors. The slack notifier only
issued log calls, but swallowed the error for handling in checkup runs.
