+++ types/errors.go
@@ -1,0 +1,29 @@
+package types
+
+import (
+	"strings"
+)
+
+// Errors is an error type that concatenates multiple errors.
+type Errors []error
+
+// Error returns a string containing all the errors in e.
+func (e Errors) Error() string {
+	var errs []string
+	for _, err := range e {
+		if err != nil {
+			errs = append(errs, err.Error())
+		}
+	}
+	return strings.Join(errs, "; ")
+}
+
+// Empty returns whether e has any non-nil errors in it.
+func (e Errors) Empty() bool {
+	for _, err := range e {
+		if err != nil {
+			return false
+		}
+	}
+	return true
+}
